(review) attach DNSDo and QUICDo to Generator
